+++ packages/breadboard/src/data/default-data-store.ts
@@ -9,6 +9,7 @@ import {
   isInlineData,
   isLLMContent,
   isStoredData,
+  LLMContent,
   SerializedDataStoreGroup,
   SerializedStoredData,
   StoredDataCapabilityPart,
@@ -18,11 +19,14 @@ import {
   asBase64,
   toStoredDataPart,
   retrieveAsBlob as genericRetrieveAsBlob,
+  isLLMContentArray,
+  isMetadataEntry,
 } from "./common.js";
 
 export type GroupID = string;
 export type NodeTimeStamp = string;
 export type OutputProperty = string;
+export type OutputPropertyIndex = number;
 export type OutputPropertyPartIndex = number;
 
 export class DefaultDataStore implements DataStore {
@@ -33,7 +37,10 @@ export class DefaultDataStore implements DataStore {
       NodeTimeStamp,
       Map<
         OutputProperty,
-        Map<OutputPropertyPartIndex, StoredDataCapabilityPart>
+        Map<
+          OutputPropertyIndex,
+          Map<OutputPropertyPartIndex, StoredDataCapabilityPart>
+        >
       >
     >
   >();
@@ -48,7 +55,10 @@ export class DefaultDataStore implements DataStore {
       NodeTimeStamp,
       Map<
         OutputProperty,
-        Map<OutputPropertyPartIndex, StoredDataCapabilityPart>
+        Map<
+          OutputPropertyIndex,
+          Map<OutputPropertyPartIndex, StoredDataCapabilityPart>
+        >
       >
     >();
     this.#dataStores.set(groupId, dataStore);
@@ -72,13 +82,19 @@ export class DefaultDataStore implements DataStore {
     if (!properties) {
       properties = new Map<
         OutputProperty,
-        Map<OutputPropertyPartIndex, StoredDataCapabilityPart>
+        Map<
+          OutputPropertyIndex,
+          Map<OutputPropertyPartIndex, StoredDataCapabilityPart>
+        >
       >();
 
       dataStore.set(nodeTimeStamp, properties);
     }
 
-    properties.set(`direct-set-${nodeTimeStamp}`, new Map([[0, part]]));
+    properties.set(
+      `direct-set-${nodeTimeStamp}`,
+      new Map([[0, new Map([[0, part]])]])
+    );
     return part;
   }
 
@@ -106,43 +122,73 @@ export class DefaultDataStore implements DataStore {
     if (!properties) {
       properties = new Map<
         OutputProperty,
-        Map<OutputPropertyPartIndex, StoredDataCapabilityPart>
+        Map<
+          OutputPropertyIndex,
+          Map<OutputPropertyPartIndex, StoredDataCapabilityPart>
+        >
       >();
       dataStore.set(nodeTimeStamp, properties);
     }
 
     for (const [property, value] of Object.entries(result.data.outputs)) {
-      if (!isLLMContent(value)) {
+      if (!isLLMContent(value) && !isLLMContentArray(value)) {
         continue;
       }
 
-      let partHandles = properties.get(property);
-      if (!partHandles) {
-        partHandles = new Map<
-          OutputPropertyPartIndex,
-          StoredDataCapabilityPart
+      const values: LLMContent[] = isLLMContent(value) ? [value] : value;
+
+      let propertyHandles = properties.get(property);
+      if (!propertyHandles) {
+        propertyHandles = new Map<
+          OutputPropertyIndex,
+          Map<OutputPropertyPartIndex, StoredDataCapabilityPart>
         >();
-        properties.set(property, partHandles);
+        properties.set(property, propertyHandles);
       }
 
-      for (let j = 0; j < value.parts.length; j++) {
-        const part = value.parts[j];
+      for (let i = 0; i < values.length; i++) {
+        const value = values[i];
+        if (isMetadataEntry(value)) {
+          continue;
+        }
 
-        if (isInlineData(part)) {
-          let storedDataPart = partHandles.get(j);
-          if (!storedDataPart) {
-            storedDataPart = await toStoredDataPart(part);
+        for (let j = 0; j < value.parts.length; j++) {
+          const part = value.parts[j];
+
+          if (isInlineData(part)) {
+            let partHandles = propertyHandles.get(0);
+            if (!partHandles) {
+              partHandles = new Map<
+                OutputPropertyIndex,
+                StoredDataCapabilityPart
+              >();
+              propertyHandles.set(i, partHandles);
+            }
+
+            let storedDataPart = partHandles.get(j);
+            if (!storedDataPart) {
+              storedDataPart = await toStoredDataPart(part);
+            }
+
+            value.parts[j] = storedDataPart;
+            partHandles.set(j, storedDataPart);
+          } else if (isStoredData(part)) {
+            let partHandles = propertyHandles.get(0);
+            if (!partHandles) {
+              partHandles = new Map<
+                OutputPropertyIndex,
+                StoredDataCapabilityPart
+              >();
+              propertyHandles.set(i, partHandles);
+            }
+
+            // This is a stored data from a previous run, so now we need to call
+            // toStoredDataPart again to create a new stored data part for this
+            // run.
+            const storedDataPart = await toStoredDataPart(part);
+            value.parts[j] = storedDataPart;
+            partHandles.set(j, storedDataPart);
           }
-
-          value.parts[j] = storedDataPart;
-          partHandles.set(j, storedDataPart);
-        } else if (isStoredData(part)) {
-          // This is a stored data from a previous run, so now we need to call
-          // toStoredDataPart again to create a new stored data part for this
-          // run.
-          const storedDataPart = await toStoredDataPart(part);
-          value.parts[j] = storedDataPart;
-          partHandles.set(j, storedDataPart);
         }
       }
     }
@@ -163,15 +209,17 @@ export class DefaultDataStore implements DataStore {
     let allHandles: Promise<SerializedStoredData>[] = [];
     for (const nodeTimeStamp of nodes.values()) {
       for (const outputProperty of nodeTimeStamp.values()) {
-        const storedData = outputProperty.values();
-        allHandles = [...storedData].map(async (storedDataPart) => {
-          const { handle } = storedDataPart.storedData;
-          const response = await fetch(handle);
-          const blob = await response.blob();
-          const mimeType = blob.type;
-          const data = await asBase64(blob);
-          return { handle, inlineData: { mimeType, data } };
-        });
+        for (const outputPropertyIndex of outputProperty.values()) {
+          const storedData = outputPropertyIndex.values();
+          allHandles = [...storedData].map(async (storedDataPart) => {
+            const { handle } = storedDataPart.storedData;
+            const response = await fetch(handle);
+            const blob = await response.blob();
+            const mimeType = blob.type;
+            const data = await asBase64(blob);
+            return { handle, inlineData: { mimeType, data } };
+          });
+        }
       }
     }
 
@@ -186,8 +234,10 @@ export class DefaultDataStore implements DataStore {
 
     for (const nodeTimeStamp of nodes.values()) {
       for (const outputProperty of nodeTimeStamp.values()) {
-        for (const storedData of outputProperty.values()) {
-          URL.revokeObjectURL(storedData.storedData.handle);
+        for (const outputPropertyIndex of outputProperty.values()) {
+          for (const storedData of outputPropertyIndex.values()) {
+            URL.revokeObjectURL(storedData.storedData.handle);
+          }
         }
       }
     }